+++ src/main/java/com/github/jlangch/venice/impl/ModuleLoader.java
@@ -102,7 +102,6 @@ public class ModuleLoader {
 							"test",
 							"tomcat", 
 							"tomcat-util", 
-							"walk",
 							"webdav", 
 							"xchart", 
 							"xml")));